v1.1.2 add new View seekbarRangeAdvance

+++ myseekbarrange/src/main/java/ru/ralnik/myseekbarrange/interfaces/OnSeekbarRangeChangeListener.java
@@ -1,5 +1,5 @@
 package ru.ralnik.myseekbarrange.interfaces;
 
 public interface OnSeekbarRangeChangeListener{
-    void SeekbarRangeValueChanged(float Thumb1Value, float Thumb2Value);
+    void SeekbarRangeValueChanged(Double Thumb1Value, Double Thumb2Value);
 }
